+++ packages/javascript/lib/string/index.ts
@@ -17,3 +17,4 @@ export { trim } from './trim.js';
 export { truncate } from './truncate.js';
 export { truncateExpect } from './truncateExpect.js';
 export { urlJoin } from './urlJoin.js';
+export { words } from './words.js';